# Management-student/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CreatePhanCongGiangDay": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"MaLopHocPhan": {
						"type": "string"
					}
				},
				"required": [
					"MaSoGiangVien",
					"MaLopHocPhan"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PhanCongGiangDay": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"MaLopHocPhan": {
						"type": "string"
					},
					"createAt": {
						"type": "string",
						"format": "date-time"
					},
					"updateAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"MaSoGiangVien",
					"MaLopHocPhan"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePhanCongGiangDay": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"MaLopHocPhan": {
						"type": "string"
					}
				},
				"required": [
					"MaSoGiangVien",
					"MaLopHocPhan"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MonHocBatBuoc": {
				"enum": [
					"Co",
					"Khong"
				],
				"type": "string"
			},
			"HinhThucHoc": {
				"enum": [
					"Online",
					"Offline"
				],
				"type": "string"
			},
			"CreateMonHoc": {
				"properties": {
					"MaMonHoc": {
						"type": "string"
					},
					"TenMonHoc": {
						"type": "string"
					},
					"BatBuoc": {
						"$ref": "#/components/schemas/MonHocBatBuoc"
					},
					"SoTinChi": {
						"type": "number",
						"format": "double"
					},
					"HocPhiMonHoc": {
						"type": "number",
						"format": "double"
					},
					"HinhThucHoc": {
						"$ref": "#/components/schemas/HinhThucHoc"
					}
				},
				"required": [
					"MaMonHoc",
					"TenMonHoc",
					"SoTinChi",
					"HocPhiMonHoc",
					"HinhThucHoc"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMonHoc": {
				"properties": {
					"MaMonHoc": {
						"type": "string"
					},
					"TenMonHoc": {
						"type": "string"
					},
					"BatBuoc": {
						"$ref": "#/components/schemas/MonHocBatBuoc"
					},
					"SoTinChi": {
						"type": "number",
						"format": "double"
					},
					"HocPhiMonHoc": {
						"type": "number",
						"format": "double"
					},
					"HinhThucHoc": {
						"$ref": "#/components/schemas/HinhThucHoc"
					}
				},
				"required": [
					"MaMonHoc",
					"TenMonHoc",
					"SoTinChi",
					"HocPhiMonHoc",
					"HinhThucHoc"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrangThaiLopHoc": {
				"enum": [
					"DangHoatDong",
					"NgungHoatDong"
				],
				"type": "string"
			},
			"CreateMaLopHocPhan": {
				"properties": {
					"MaLopHocPhan": {
						"type": "string"
					},
					"MaMonHoc": {
						"type": "string"
					},
					"NgayBatDau": {
						"type": "string",
						"format": "date-time"
					},
					"NgayKetThuc": {
						"type": "string",
						"format": "date-time"
					},
					"SoSinhVienToiThieu": {
						"type": "number",
						"format": "double"
					},
					"SoSinhVienToiDa": {
						"type": "number",
						"format": "double"
					},
					"SoSinhVienHienTai": {
						"type": "number",
						"format": "double"
					},
					"TrangThaiLopHoc": {
						"$ref": "#/components/schemas/TrangThaiLopHoc"
					}
				},
				"required": [
					"MaLopHocPhan",
					"MaMonHoc",
					"NgayBatDau",
					"NgayKetThuc",
					"SoSinhVienToiThieu",
					"SoSinhVienToiDa",
					"SoSinhVienHienTai",
					"TrangThaiLopHoc"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMaLopHocPhan": {
				"properties": {
					"MaLopHocPhan": {
						"type": "string"
					},
					"MaMonHoc": {
						"type": "string"
					},
					"NgayBatDau": {
						"type": "string",
						"format": "date-time"
					},
					"NgayKetThuc": {
						"type": "string",
						"format": "date-time"
					},
					"SoSinhVienToiThieu": {
						"type": "number",
						"format": "double"
					},
					"SoSinhVienToiDa": {
						"type": "number",
						"format": "double"
					},
					"SoSinhVienHienTai": {
						"type": "number",
						"format": "double"
					},
					"TrangThaiLopHoc": {
						"$ref": "#/components/schemas/TrangThaiLopHoc"
					}
				},
				"required": [
					"MaLopHocPhan",
					"MaMonHoc",
					"NgayBatDau",
					"NgayKetThuc",
					"SoSinhVienToiThieu",
					"SoSinhVienToiDa",
					"SoSinhVienHienTai",
					"TrangThaiLopHoc"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "management-student",
		"version": "0.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/API/v1/PhanCongGiangDay/CreatePhanCongGiangDay": {
			"post": {
				"operationId": "CreatePhanCongGiangDay",
				"responses": {
					"201": {
						"description": "Tạo thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/CreatePhanCongGiangDay"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PhanCongGiangDay"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePhanCongGiangDay"
							}
						}
					}
				}
			}
		},
		"/API/v1/PhanCongGiangDay/TimkiemPhanCongGiangDay/{MaSoGiangVien}/{MaLopHocPhan}": {
			"get": {
				"operationId": "TimkiemPhanCongGiangDay",
				"responses": {
					"200": {
						"description": "Tìm kiếm thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/PhanCongGiangDay"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PhanCongGiangDay"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/PhanCongGiangDay/LietKePhanCongHocPhan": {
			"get": {
				"operationId": "LietKePhanCongHocPhan",
				"responses": {
					"200": {
						"description": "Liệt kê thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"PhanCongGiangDay"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/PhanCongGiangDay/CapNhapLopHocPhan/{MaLopHocPhan}": {
			"put": {
				"operationId": "CapNhapLopHocPhan",
				"responses": {
					"200": {
						"description": "Cập nhập thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PhanCongGiangDay"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePhanCongGiangDay"
							}
						}
					}
				}
			}
		},
		"/API/v1/PhanCongGiangDay/XoaLopHocPhan/{MaSoGiangVien}/{MaLopHocPhan}": {
			"delete": {
				"operationId": "XoaLopHocPhan",
				"responses": {
					"200": {
						"description": "Xoá thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PhanCongGiangDay"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/MonHoc/CreateMonHoc": {
			"post": {
				"operationId": "CreateMonHoc",
				"responses": {
					"201": {
						"description": "Tạo thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MonHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMonHoc"
							}
						}
					}
				}
			}
		},
		"/API/v1/MonHoc/GetOneMonHoc/{MaMonHoc}": {
			"get": {
				"operationId": "GetOneMonHoc",
				"responses": {
					"200": {
						"description": "Tìm kiếm thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MonHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/MonHoc/GetAllMonHoc": {
			"get": {
				"operationId": "GetAllMonHoc",
				"responses": {
					"200": {
						"description": "Liệt kê thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MonHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/MonHoc/CapNhapMonHoc/{MaMonHoc}": {
			"put": {
				"operationId": "CapNhapMonHoc",
				"responses": {
					"200": {
						"description": "Cập nhập thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MonHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMonHoc"
							}
						}
					}
				}
			}
		},
		"/API/v1/MonHoc/DeleteMonHoc/{MaMonHoc}": {
			"delete": {
				"operationId": "DeleteMonHoc",
				"responses": {
					"200": {
						"description": "Xoá thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"MonHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/LopHoc/CreateLopHoc": {
			"post": {
				"operationId": "CreateLopHoc",
				"responses": {
					"201": {
						"description": "Tạo thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LopHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMaLopHocPhan"
							}
						}
					}
				}
			}
		},
		"/API/v1/LopHoc/GetOneLopHoc/{MaLopHocPhan}": {
			"get": {
				"operationId": "GetOneLopHoc",
				"responses": {
					"200": {
						"description": "Tìm kiếm thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LopHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/LopHoc/GetAllLopHoc": {
			"get": {
				"operationId": "GetAllLopHoc",
				"responses": {
					"200": {
						"description": "Liệt kê thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LopHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/LopHoc/UpdateLopHoc/{MaLopHocPhan}": {
			"put": {
				"operationId": "UpdateMonHoc",
				"responses": {
					"200": {
						"description": "Cập nhập thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LopHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMaLopHocPhan"
							}
						}
					}
				}
			}
		},
		"/API/v1/LopHoc/DeleteMonHoc/{MaLopHocPhan}": {
			"delete": {
				"operationId": "DeleteMonHoc",
				"responses": {
					"200": {
						"description": "Xoá thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LopHoc"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}